requirements.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.5.7
click==8.1.6
colorama==0.4.6
distlib==0.3.7
dnspython==2.4.0
email-validator==2.0.0.post2
fastapi==0.100.0
filelock==3.12.2
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
orjson==3.9.2
platformdirs==3.9.1
pydantic==2.0.3
pydantic-extra-types==2.0.0
pydantic-settings==2.0.2
pydantic_core==2.3.0
pymssql==2.2.7
pyodbc==4.0.34
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
ujson==5.8.0
uvicorn==0.23.1
virtualenv==20.24.1
watchfiles==0.19.0
websockets==11.0.3
